project done :)

+++ hello_world/6-size.c
@@ -7,11 +7,11 @@
 
 int main(void)
 {
-	printf("Size of a char: %lu bytes\n", (unsigned long) sizeof(char));
-	printf("size of an int: %lu bytes\n", (unsigned long) sizeof(int));
-	printf("size of a long int: %lu bytes\n", (unsigned long) sizeof(long int));
-	printf("size of a long long int: %lu bytes\n", (unsigned long) sizeof(long long int));
-	printf("size of float: %lu bytes\n", (unsigned long) sizeof(float));
-	
+	printf("Size of a char: %zu bytes\n", sizeof(char));
+	printf("size of an int: %zu bytes\n", sizeof(int));
+	printf("size of a long int: %zu bytes\n", sizeof(long int));
+	printf("size of a long long int: %zu bytes\n", sizeof(long long int));
+	printf("size of float: %zu bytes\n", sizeof(float));
+
 		return (0);
 }
